+++ pkg/common/oras/authprovider/azure/azureworkloadidentity.go
@@ -41,34 +41,6 @@ type WIAuthProvider struct {
 	reportMetrics     func(ctx context.Context, duration int64, artifactHostName string)
 }
 
-func defaultAuthClientFactory(serverURL string, options *azcontainerregistry.AuthenticationClientOptions) (AuthClient, error) {
-	client, err := azcontainerregistry.NewAuthenticationClient(serverURL, options)
-	if err != nil {
-		return nil, err
-	}
-	return &AuthenticationClientWrapper{client: client}, nil
-}
-
-func defaultGetAADAccessToken(ctx context.Context, tenantID, clientID, resource string) (confidential.AuthResult, error) {
-	return azureauth.GetAADAccessToken(ctx, tenantID, clientID, resource)
-}
-
-func defaultReportMetrics(ctx context.Context, duration int64, artifactHostName string) {
-	logger.GetLogger(ctx, logOpt).Infof("Metrics Report: Duration=%dms, Host=%s", duration, artifactHostName)
-}
-
-type AuthenticationClientWrapper struct {
-	client *azcontainerregistry.AuthenticationClient
-}
-
-func (w *AuthenticationClientWrapper) ExchangeAADAccessTokenForACRRefreshToken(ctx context.Context, grantType, service string, options *azcontainerregistry.AuthenticationClientExchangeAADAccessTokenForACRRefreshTokenOptions) (azcontainerregistry.AuthenticationClientExchangeAADAccessTokenForACRRefreshTokenResponse, error) {
-	return w.client.ExchangeAADAccessTokenForACRRefreshToken(ctx, azcontainerregistry.PostContentSchemaGrantType(grantType), service, options)
-}
-
-type AuthClient interface {
-	ExchangeAADAccessTokenForACRRefreshToken(ctx context.Context, grantType, service string, options *azcontainerregistry.AuthenticationClientExchangeAADAccessTokenForACRRefreshTokenOptions) (azcontainerregistry.AuthenticationClientExchangeAADAccessTokenForACRRefreshTokenResponse, error)
-}
-
 type azureWIAuthProviderConf struct {
 	Name     string `json:"name"`
 	ClientID string `json:"clientID,omitempty"`
@@ -118,9 +90,9 @@ func (s *AzureWIProviderFactory) Create(authProviderConfig provider.AuthProvider
 		aadToken:          token,
 		tenantID:          tenant,
 		clientID:          clientID,
-		authClientFactory: defaultAuthClientFactory,
-		getAADAccessToken: defaultGetAADAccessToken,
-		reportMetrics:     defaultReportMetrics,
+		authClientFactory: DefaultAuthClientFactory,
+		getAADAccessToken: DefaultGetAADAccessToken,
+		reportMetrics:     DefaultReportMetrics,
 	}, nil
 }
